Support for dependencies
Brunch won't recompile tag file if it has any include/import-d
files and they are changed. For example, if Jade is used for
template, and it `include`s LESS file or CoffeeScript file,
making change to those files will not trigger Brunch's build.

This commit uses `progeny` to retrieve those dependencies from
the source file and pass to Brunch to keep track of.