require and export in CommonJS style
riot-brunch assumes that we have the `riot` global available, meaning I have to add `riot.js` as a vendor script.  This is unnecessary as we're downloading it from the NPM anyway and it can be included with a simple `require`.  This way also means polluting the global scope.

This change does things in what I feel is the "CommonJS way".  The code comes from https://github.com/jhthorsen/riotify/blob/master/index.js#L16 .  We eliminate the vendor script and the global.