Add 'requireRiot' configuration parameter.
Right now when brunch wraps riot-generated code
into module, that module does not have a 'riot' symbol
defined. This causes runtime errors in the browser
when the tag is loaded/imported.

If the 'requireRiot' parameter is set to 'true'
following statement will be prepended to
the compiled template:

  var riot = require('riot');

This way, when the brunch wraps the tag code into
isolated module, 'riot' symbol will be defined
inside of that module.